修正RGBLIGHT_EFFECT_TWINKLE无法工作问题

+++ application/main/src/driver/rgb_matrix/rgblight.c
@@ -1405,7 +1405,7 @@ void rgblight_effect_twinkle(animation_status_t* anim)
     const uint8_t top = breathe_calc(127);
 
     uint8_t frac(uint8_t n, uint8_t d) { return (uint16_t)255 * n / d; }
-    uint8_t scale(uint16_t v, uint8_t scale) { return (v * scale) >> 8; }
+    uint16_t scale(uint16_t v, uint8_t scale) { return ((uint32_t)v * scale) >> 8; }
 
     for (uint8_t i = 0; i < rgblight_ranges.effect_num_leds; i++) {
         TwinkleState* t = &(led_twinkle_state[i]);
@@ -1425,7 +1425,7 @@ void rgblight_effect_twinkle(animation_status_t* anim)
             t->life--;
             uint8_t unscaled = frac(breathe_calc(frac(t->life, t->max_life)) - bottom, top - bottom);
             c->v = scale(rgblight_config.val, unscaled);
-        } else if (rand() < scale((uint16_t)RAND_MAX * RGBLIGHT_EFFECT_TWINKLE_PROBABILITY, 127 + rgblight_config.val / 2)) {
+        } else if ((uint16_t)rand() < scale((uint16_t)RAND_MAX * RGBLIGHT_EFFECT_TWINKLE_PROBABILITY, 127 + rgblight_config.val / 2)) {
             // This LED is off, but was randomly selected to start brightening
             if (random_color) {
                 c->h = rand() % 0xFF;
